- Add Redux - Add Search

+++ src/app/components/post-card/post-card.component.ts
@@ -7,6 +7,11 @@ import { PostsService } from 'src/app/services/posts.service';
 
 //Map
 import * as  mapboxgl from 'mapbox-gl';
+import { Store } from '@ngrx/store';
+import { AppState } from 'src/app/app.reducers';
+import { DeletePost } from 'src/app/redux/actions/post.actions';
+import { MatDialog } from '@angular/material/dialog';
+import { ModalConfirmComponent } from '../shared/modal-confirm/modal-confirm.component';
 
 @Component({
   selector: 'app-post-card',
@@ -25,7 +30,9 @@ export class PostCardComponent implements OnInit, AfterViewInit {
   constructor(
     private _postsService: PostsService,
     private _snackBar: MatSnackBar,
-    private router: Router
+    private router: Router,
+    private store: Store<AppState>,
+    private dialog: MatDialog,
   ) { }
 
   ngOnInit(): void {
@@ -49,7 +56,6 @@ export class PostCardComponent implements OnInit, AfterViewInit {
       const maker = new mapboxgl.Marker()
         .setLngLat([this.post?.long, this.post?.lat])
         .addTo(this.map)
-
     }
   }
 
@@ -58,13 +64,23 @@ export class PostCardComponent implements OnInit, AfterViewInit {
   }
 
   delete() {
-    this._postsService.deletePost(this.post?.id).subscribe(
-      () => {
-        this._snackBar.open('Post Deleted', 'Done', {
-          duration: 3000
-        });
-        this.router.navigate([''])
+    this.dialog.open(ModalConfirmComponent, {
+      data: {
+        title: this.post.title,
       }
-    )
+    }).afterClosed().subscribe(result => {
+      if (result) {
+        this._postsService.deletePost(this.post?.id).subscribe(
+          () => {
+            const action = new DeletePost(this.post?.id);
+            this.store.dispatch(action);
+            this._snackBar.open('Post Deleted', 'Done', {
+              duration: 3000
+            });
+            this.router.navigate([''])
+          }
+        )
+      }
+    });
   }
 }
